⚡ Make python nedryland 9.0.0 compatible
With the new version of Nedryland, mkShell is no longer used, which
means that inputs are not resolved before the shell is created. Now we
need to place the checkHook first manually. And place it in
passthru.shellInputs.

TargetSetup has moved to shellInputs. This is because the target
setup is now a setup hook instead of something that the nedryland
shell does.