🐛 Fix & Improve python check runner
There are cases where TMP env is not set which made some tools not
pick up configuration. Using mktemp and generating a new file every
time instead.

In the case where you were in a shell and changed your directory the
script would also miss the configuration files. Fixed so it always
looks in component dir.