⚡ Add input and remove targetName
Add input as a parameter to `mkPlatform` for C. This makes it
possible to use a different component target when components are
listed as dependencies in other components.

Remove `targetName` as a parameter sent to the `mk*` functions in C.
`targetName` is only a string and does not represent the expected
information since it can sometimes refer to what is being used from
other components and does not reflect any state on the current
component. Instead, accept `stdenv` as a parameter and use that one.

Also fix so that the spliced stdenv is the stdenv actually used.