⚡ 🐍 Change python default to pyproject.toml
The days of setuptools are numbered! Switch default component to use
pyproject.toml, it is currently the least deprecated way of building
python packages.

Also sneak in some nifty features like adding a `run` command for
clients and services. And add a corresponding entry point to their
package.

Also made the generated code pass `ruff`, it uses the
`--unsafe-fixes` to add period at the end of docstrings if missing.
Since it only runs on the templated code we know that it won't do
anything else.