⚡ Add wrapped grpcurl to protobuf's mkModule
grpcurl is a curl like CLI for serializing and deserializing command
line arguments to protobuf, which makes it very useful while developing
a grpc endpoint. However it requires some generated files to be
specified and those can be a bit tricky to locate in the nix store so we
wrap the exe with the files.