✨ Add overrideAttrsPre to C
This function can be used when overriding `mkDerivation` for the C
language and has two arguments instead of one as `overrideAttrs`.

The first argument gives access to the attributes sent in to
`mkDerivation` by the user, before anything has been modified. The
second argument are the attributes after modifications, before
sending to the actual `mkDerivation`.